requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
click==8.1.8
    # via uvicorn
cramjam==2.10.0
    # via fastparquet
exceptiongroup==1.3.0
    # via anyio
fastapi==0.115.12
    # via -r requirements.in
fastparquet==2024.11.0
    # via -r requirements.in
fsspec==2025.5.1
    # via fastparquet
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
lightgbm==4.6.0
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   fastparquet
    #   lightgbm
    #   pandas
    #   pyarrow
    #   scipy
    #   xgboost
packaging==25.0
    # via fastparquet
pandas==2.2.2
    # via
    #   -r requirements.in
    #   fastparquet
pyarrow==12.0.1
    # via -r requirements.in
pydantic==2.11.5
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   pandas
pytz==2025.2
    # via
    #   -r requirements.in
    #   pandas
scipy==1.13.1
    # via
    #   -r requirements.in
    #   lightgbm
    #   xgboost
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uvicorn==0.34.2
    # via -r requirements.in
xgboost==2.1.4
    # via -r requirements.in
# pyarrow==12.0.1
google-cloud-storage>=2.0.0